requirements.txt
google-genai
spacy
datasets
google-cloud-aiplatform
